+++ pandora/img_tools.py
@@ -294,7 +294,7 @@ def create_dataset_from_inputs(input_config: dict, roi: dict = None) -> xr.Datas
     :return: xarray.DataSet containing the variables :
 
             - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-            - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+            - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
             - msk (optional): 2D (row, col) xarray.DataArray int16
             - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
             - segm (optional): 2D (row, col) xarray.DataArray int16
