+++ pandora/matching_cost/census.py
@@ -83,7 +83,7 @@ class Census(matching_cost.AbstractMatchingCost):
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -91,7 +91,7 @@ class Census(matching_cost.AbstractMatchingCost):
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -203,7 +203,7 @@ class Census(matching_cost.AbstractMatchingCost):
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -211,7 +211,7 @@ class Census(matching_cost.AbstractMatchingCost):
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
