+++ pandora/matching_cost/matching_cost.py
@@ -172,7 +172,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -180,7 +180,7 @@ class AbstractMatchingCost:
         :param img_right: right Dataset  containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -224,7 +224,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -232,7 +232,7 @@ class AbstractMatchingCost:
         :param img_right: right Dataset  containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -263,7 +263,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -342,7 +342,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -350,7 +350,7 @@ class AbstractMatchingCost:
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -394,7 +394,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -402,7 +402,7 @@ class AbstractMatchingCost:
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -528,7 +528,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -536,7 +536,7 @@ class AbstractMatchingCost:
         :param img_right: right Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
@@ -609,7 +609,7 @@ class AbstractMatchingCost:
         :param img_left: left Dataset image containing :
 
                 - im: 2D (row, col) or 3D (band_im, row, col) xarray.DataArray float32
-                - disparity (optional): 3D (disp, row, col) xarray.DataArray int16
+                - disparity (optional): 3D (disp, row, col) xarray.DataArray float32
                 - msk (optional): 2D (row, col) xarray.DataArray int16
                 - classif (optional): 3D (band_classif, row, col) xarray.DataArray int16
                 - segm (optional): 2D (row, col) xarray.DataArray int16
